Move DirectX wrappers to new project
