+++ src/BizHawk.Client.EmuHawk/GraphicsImplementations/IGLExtensions.cs
@@ -1,6 +1,7 @@
 using System;
 
 using BizHawk.Bizware.BizwareGL;
+using BizHawk.Bizware.DirectX;
 
 namespace BizHawk.Client.EmuHawk
 {
