+++ src/BizHawk.Client.EmuHawk/Program.cs
@@ -6,6 +6,8 @@ using System.Reflection;
 using System.Runtime.InteropServices;
 using System.Windows.Forms;
 
+using BizHawk.Bizware.DirectX;
+
 using Microsoft.VisualBasic.ApplicationServices;
 
 using BizHawk.Common;
