+++ src/BizHawk.Client.EmuHawk/config/SoundConfig.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Windows.Forms;
 
+using BizHawk.Bizware.DirectX;
 using BizHawk.Client.Common;
 using BizHawk.Common;
